Add files via upload
The C program initializes a character variable c with the value 'Z' and declares a pointer to a character cp pointing to the address of c. The program prints the initial address and the value of the character pointed to by cp. It then employs pointer arithmetic by incrementing the pointer cp by 1, reflecting the size of a char. The program prints the updated address of cp, demonstrating the effect of pointer arithmetic. However, it refrains from printing the value of *cp after the increment, cautioning that the pointer might now point to memory not allocated for the program. The code serves as a concise illustration of how pointer arithmetic works with a char variable and encourages exploration of similar operations with different data types, such as int and double, to understand their respective sizes in memory.